Rearchs provider config
