+++ commands/login_test.go
@@ -74,14 +74,6 @@ func Mocks(t *testing.T) (*pktoken.PKToken, crypto.Signer, providers.OpenIdProvi
 	return pkt, signer, op
 }
 
-func ProviderFromString(t *testing.T, providerString string) providers.OpenIdProvider {
-	providerConfig3, err := NewProviderConfigFromString(providerStr3, true)
-	require.NoError(t, err)
-	provider3, err := NewProviderFromConfig(providerConfig3, false)
-	require.NoError(t, err)
-	return provider3
-}
-
 func TestLoginCmd(t *testing.T) {
 	_, _, mockOp := Mocks(t)
 
@@ -159,7 +151,7 @@ func TestDetermineProvider(t *testing.T) {
 			wantIssuer:    "",
 			wantError:     false,
 			errorString:   "",
-			wantChooser:   `[{"ClientSecret":"","Scopes":[""],"PromptType":"consent","AccessType":"offline","RedirectURIs":["http://localhost:3000/login-callback","http://localhost:10001/login-callback","http://localhost:11110/login-callback"],"GQSign":false,"OpenBrowser":false,"HttpClient":null,"IssuedAtOffset":60000000000}]`,
+			wantChooser:   `[{"ClientSecret":"","Scopes":["openid profile email"],"PromptType":"consent","AccessType":"offline","RedirectURIs":["http://localhost:3000/login-callback","http://localhost:10001/login-callback","http://localhost:11110/login-callback"],"GQSign":false,"OpenBrowser":false,"HttpClient":null,"IssuedAtOffset":60000000000}]`,
 		},
 		{
 			name:          "Good path with env vars many providers and no default",
@@ -168,7 +160,7 @@ func TestDetermineProvider(t *testing.T) {
 			providerAlias: "",
 			wantIssuer:    "",
 			wantError:     false,
-			wantChooser:   `[{"ClientSecret":"","Scopes":[""],"PromptType":"consent","AccessType":"offline","RedirectURIs":["http://localhost:3000/login-callback","http://localhost:10001/login-callback","http://localhost:11110/login-callback"],"GQSign":false,"OpenBrowser":false,"HttpClient":null,"IssuedAtOffset":60000000000},{"ClientSecret":"","Scopes":[""],"PromptType":"consent","AccessType":"offline","RedirectURIs":["http://localhost:3000/login-callback","http://localhost:10001/login-callback","http://localhost:11110/login-callback"],"GQSign":false,"OpenBrowser":false,"HttpClient":null,"IssuedAtOffset":60000000000},{"ClientSecret":"","Scopes":[""],"PromptType":"consent","AccessType":"offline","RedirectURIs":["http://localhost:3000/login-callback","http://localhost:10001/login-callback","http://localhost:11110/login-callback"],"GQSign":false,"OpenBrowser":false,"HttpClient":null,"IssuedAtOffset":60000000000}]`,
+			wantChooser:   `[{"ClientSecret":"","Scopes":["openid profile email"],"PromptType":"consent","AccessType":"offline","RedirectURIs":["http://localhost:3000/login-callback","http://localhost:10001/login-callback","http://localhost:11110/login-callback"],"GQSign":false,"OpenBrowser":false,"HttpClient":null,"IssuedAtOffset":60000000000},{"ClientSecret":"","Scopes":["openid profile email"],"PromptType":"consent","AccessType":"offline","RedirectURIs":["http://localhost:3000/login-callback","http://localhost:10001/login-callback","http://localhost:11110/login-callback"],"GQSign":false,"OpenBrowser":false,"HttpClient":null,"IssuedAtOffset":60000000000},{"ClientSecret":"","Scopes":["openid profile email"],"PromptType":"consent","AccessType":"offline","RedirectURIs":["http://localhost:3000/login-callback","http://localhost:10001/login-callback","http://localhost:11110/login-callback"],"GQSign":false,"OpenBrowser":false,"HttpClient":null,"IssuedAtOffset":60000000000}]`,
 		},
 		{
 			name:          "Good path with env vars many providers and providerAlias",
@@ -239,99 +231,6 @@ func TestDetermineProvider(t *testing.T) {
 			}
 		})
 	}
-}
-
-func TestProviderConfigFromString(t *testing.T) {
-
-	tests := []struct {
-		name           string
-		configString   string
-		hasAlias       bool
-		expectedIssuer string
-		wantError1     bool
-		errorString1   string
-		wantError2     bool
-		errorString2   string
-	}{
-		{
-			name:           "Good path with test providerStr3",
-			configString:   providerStr3,
-			hasAlias:       true,
-			expectedIssuer: providerIssuer3,
-		},
-		{
-			name:           "Good path with test authentik OP",
-			configString:   "authentik,https://authentik.io/application/o/opkssh/,client_id,,openid profile email",
-			hasAlias:       true,
-			expectedIssuer: "https://authentik.io/application/o/opkssh/",
-		},
-		{
-			name:           "Good path with test Google OP",
-			configString:   "https://accounts.google.com,206584157355-7cbe4s640tvm7naoludob4ut1emii7sf.apps.googleusercontent.com,NOT-aREAL_3a_GOOGLE-CLIENTSECRET",
-			hasAlias:       false,
-			expectedIssuer: "https://accounts.google.com",
-		},
-		{
-			name:           "Good path with test microsoft OP",
-			configString:   "https://login.microsoftonline.com/9188040d-6c67-4c5b-b112-36a304b66dad/v2.0,096ce0a3-5e72-4da8-9c86-12924b294a01",
-			hasAlias:       false,
-			expectedIssuer: "https://login.microsoftonline.com/9188040d-6c67-4c5b-b112-36a304b66dad/v2.0",
-		},
-		{
-			name:           "Good path with test microsoft OP",
-			configString:   "https://gitlab.com,8d8b7024572c7fd501f64374dec6bba37096783dfcd792b3988104be08cb6923",
-			hasAlias:       false,
-			expectedIssuer: "https://gitlab.com",
-		},
-		{
-			name:           "Good path with test hello OP",
-			configString:   "https://issuer.hello.coop,client-id,,openid email",
-			hasAlias:       false,
-			expectedIssuer: "https://issuer.hello.coop",
-		},
-		{
-			name:           "Alias set but no alias expected",
-			configString:   "exampleOp,https://token.example.com/,client_id,,openid profile email,",
-			hasAlias:       false,
-			expectedIssuer: "https://token.example.com/",
-			wantError2:     true,
-			errorString2:   "invalid provider issuer value. Expected issuer to start with 'https://'",
-		},
-		{
-			name:           "No alias set but alias expected",
-			configString:   "https://token.example.com/,client_id,,openid profile email,",
-			hasAlias:       true,
-			expectedIssuer: "https://token.example.com/",
-			wantError1:     true,
-			errorString1:   "invalid provider client-ID value got ()",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			providerConfig, err := NewProviderConfigFromString(tt.configString, tt.hasAlias)
-			if tt.wantError1 {
-				require.Error(t, err, "Expected error but got none")
-				if tt.errorString1 != "" {
-					require.ErrorContains(t, err, tt.errorString1, "Got a wrong error message")
-				}
-
-			} else {
-				require.NoError(t, err)
-				provider, err := NewProviderFromConfig(providerConfig, false)
-				if tt.wantError2 {
-					require.Error(t, err, "Expected error but got none")
-					if tt.errorString2 != "" {
-						require.ErrorContains(t, err, tt.errorString2, "Got a wrong error message")
-					}
-				} else {
-					require.NoError(t, err)
-					require.Equal(t, tt.expectedIssuer, provider.Issuer())
-				}
-			}
-		})
-	}
-
 }
 
 func TestNewLogin(t *testing.T) {
